requirements.txt
numpy>=1.26,<3.0
pandas>=2.2,<3.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.8,<3.0
httpx>=0.28,<1.0
torch>=2.3,<3.0
torch-geometric>=2.6,<3.0
sentence-transformers>=3.0,<4.0
faiss-cpu>=1.8,<2.0
lightgbm>=4.5,<5.0
mlflow>=2.17,<3.0
dvc>=3.59,<4.0

